+++ robotidy/api.py
@@ -42,6 +42,7 @@ class RobotidyAPI(Robotidy):
             check=False,
             output=output,
             force_order=False,
+            color=False,
         )