+++ robotidy/cli.py
@@ -1,3 +1,4 @@
+import os
 import re
 from pathlib import Path
 from typing import Any, Dict, Iterable, List, Optional, Pattern, Tuple
@@ -229,6 +230,12 @@ def print_transformers_list():
     show_default=True,
 )
 @click.option(
+    "--color/--no-color",
+    default=True,
+    help="Enable ANSI coloring the output",
+    show_default=True,
+)
+@click.option(
     "--check",
     is_flag=True,
     help="Don't overwrite files and just return status. Return code 0 means nothing would change. "
@@ -328,6 +335,7 @@ def cli(
     extend_exclude: Optional[Pattern],
     overwrite: bool,
     diff: bool,
+    color: bool,
     check: bool,
     spacecount: int,
     lineseparator: str,
@@ -365,6 +373,9 @@ def cli(
         # None is default, with check not set -> overwrite, with check set -> overwrite only when overwrite flag is set
         overwrite = not check
 
+    if color:
+        color = "NO_COLOR" not in os.environ
+
     formatting_config = GlobalFormattingConfig(
         space_count=spacecount,
         line_sep=lineseparator,
@@ -386,6 +397,7 @@ def cli(
         check=check,
         output=output,
         force_order=force_order,
+        color=color,
     )
     status = tidy.transform_files()
     ctx.exit(status)